requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
pandas==1.5.3
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic-settings==2.0.3
cloudinary==1.36.0
